PAL-619  Refactor existing Policy integration tests are are not working
